refactor(algorithm): 将evaluate方法返回类型从int64_t改为double
- 所有算法实现类中的evaluate方法返回类型从int64_t更改为double
- 这包括DiffFirstAlgorithm、IDAStarAlgorithm、
  DynamicPenaltyBalancingAlgorithm、AStarAlgorithm、
  DFSAlgorithm、DPMergeAlgorithm、HammingAlgorithm等
- 更新了IAlgorithm接口中evaluate方法的声明
- 修改了测试代码中的TestAlgorithm和SlowAlgorithm实现
- AStarAlgorithm和DPMergeAlgorithm中的数值计算逻辑也相应调整
 以适配新的double类型返回值

+++ plugins/diff_first/DiffFirstAlgorithm.cpp
@@ -212,7 +212,7 @@ bool DiffFirstAlgorithm::simulate(const AlgorithmInput& input) const noexcept {
 
 // ─── evaluate ──────────────────────────────────────────────────────────────────
 
-int64_t DiffFirstAlgorithm::evaluate(int16_t ench_count) const noexcept {
+double DiffFirstAlgorithm::evaluate(int16_t ench_count) const noexcept {
     (void)ench_count;
     // O(n²) deterministic construction — sub-millisecond.
     return 0;
