+++ src/domain/algorithm/_strategies/dfs/DFSAlgorithm.h
@@ -19,7 +19,7 @@ class DFSAlgorithm : public IAlgorithm {
 
     std::string_view name() const noexcept override { return "dfs"; }
     std::string_view version() const noexcept override { return "1.0.0"; }
-    int64_t evaluate(int16_t ench_count) const noexcept override;
+    double evaluate(int16_t ench_count) const noexcept override;
     void execute(const AlgorithmInput &input, ExecutionContext &ctx) override;
     std::optional<Item> process(const EnchSolution &solution) const override;
     std::unique_ptr<IForgeEngine> get_forge_engine() const noexcept override {
