+++ tests/integration/test_algorithm_executor.cpp
@@ -32,7 +32,7 @@ class TestAlgorithm : public IAlgorithm {
 public:
     std::string_view name() const noexcept override { return "test"; }
     std::string_view version() const noexcept override { return "1.0.0"; }
-    int64_t evaluate(int16_t) const noexcept override { return 0; }
+    double evaluate(int16_t) const noexcept override { return 0; }
     std::optional<Item> process(const EnchSolution &) const override { return std::nullopt; }
     std::unique_ptr<IForgeEngine> get_forge_engine() const noexcept override {
         return std::make_unique<TestForgeEngine>();
@@ -55,7 +55,7 @@ class SlowAlgorithm : public IAlgorithm {
 public:
     std::string_view name() const noexcept override { return "slow"; }
     std::string_view version() const noexcept override { return "1.0.0"; }
-    int64_t evaluate(int16_t) const noexcept override { return 0; }
+    double evaluate(int16_t) const noexcept override { return 0; }
     std::optional<Item> process(const EnchSolution &) const override { return std::nullopt; }
     std::unique_ptr<IForgeEngine> get_forge_engine() const noexcept override {
         return std::make_unique<TestForgeEngine>();
@@ -76,7 +76,7 @@ class ThrowingAlgorithm : public IAlgorithm {
 public:
     std::string_view name() const noexcept override { return "throwing"; }
     std::string_view version() const noexcept override { return "1.0.0"; }
-    int64_t evaluate(int16_t) const noexcept override { return 0; }
+    double evaluate(int16_t) const noexcept override { return 0; }
     std::optional<Item> process(const EnchSolution &) const override { return std::nullopt; }
     std::unique_ptr<IForgeEngine> get_forge_engine() const noexcept override {
         return std::make_unique<TestForgeEngine>();
